add helpers
format time in format XX min XX sec
